Minor changes
git-svn-id: http://svn.castleproject.org:8080/svn/castle/trunk@1530 73e77b4c-caa6-f847-a29a-24ab75ae54b6

+++ Samples/MonoRail/BinderSample/BinderSample.Web/schema.sql
@@ -1,6 +1,6 @@
 alter table Book  drop constraint FK7C80C0CCEB61E439
-drop table Publisher
-drop table Book
+if exists (select * from dbo.sysobjects where id = object_id(N'Publisher') and OBJECTPROPERTY(id, N'IsUserTable') = 1) drop table Publisher
+if exists (select * from dbo.sysobjects where id = object_id(N'Book') and OBJECTPROPERTY(id, N'IsUserTable') = 1) drop table Book
 create table Publisher (
   Id INT IDENTITY NOT NULL,
    Name NVARCHAR(255) null,
